perf: additional drawing functions moved to a separate file
